# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceCircle API",
    "description": "Backend for VoiceCircle: voice enrollment, practice scam calls, deepfake checks and Hugh companion calls.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/practice/scenarios": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_api_v1_practice_scenarios_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioOut"
                  },
                  "title": "Response List Scenarios Api V1 Practice Scenarios Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/circles/{circle_id}/practice-calls": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Create Practice Call",
        "operationId": "create_practice_call_api_v1_circles__circle_id__practice_calls_post",
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeCallIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeCallOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "List Practice Calls",
        "operationId": "list_practice_calls_api_v1_circles__circle_id__practice_calls_get",
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PracticeCallOut"
                  },
                  "title": "Response List Practice Calls Api V1 Circles  Circle Id  Practice Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-calls/{call_id}": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Get Practice Call",
        "operationId": "get_practice_call_api_v1_practice_calls__call_id__get",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeCallOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-calls/{call_id}/cancel": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Cancel Practice Call",
        "operationId": "cancel_practice_call_api_v1_practice_calls__call_id__cancel_post",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeCallOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/circles/{circle_id}/detections": {
      "post": {
        "tags": [
          "detection"
        ],
        "summary": "Create Detection",
        "operationId": "create_detection_api_v1_circles__circle_id__detections_post",
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_detection_api_v1_circles__circle_id__detections_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "detection"
        ],
        "summary": "List Detections",
        "operationId": "list_detections_api_v1_circles__circle_id__detections_get",
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DetectionOut"
                  },
                  "title": "Response List Detections Api V1 Circles  Circle Id  Detections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/detections/{detection_id}": {
      "get": {
        "tags": [
          "detection"
        ],
        "summary": "Get Detection",
        "operationId": "get_detection_api_v1_detections__detection_id__get",
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Detection Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_detection_api_v1_circles__circle_id__detections_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "claimed_member_id": {
            "type": "string",
            "title": "Claimed Member Id"
          }
        },
        "type": "object",
        "required": [
          "file",
          "claimed_member_id"
        ],
        "title": "Body_create_detection_api_v1_circles__circle_id__detections_post"
      },
...
      "DetectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "circle_id": {
            "type": "string",
            "title": "Circle Id"
          },
          "claimed_member_id": {
            "type": "string",
            "title": "Claimed Member Id"
          },
          "claimed_member_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed Member Name"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "synthetic_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthetic Score"
          },
          "speaker_match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Match Score"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "circle_id",
          "claimed_member_id",
          "original_filename",
          "status",
          "verdict",
          "synthetic_score",
          "speaker_match_score",
          "explanation",
          "error",
          "created_at"
        ],
        "title": "DetectionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PracticeCallIn": {
        "properties": {
          "senior_member_id": {
            "type": "string",
            "title": "Senior Member Id"
          },
          "voice_member_id": {
            "type": "string",
            "title": "Voice Member Id"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ],
            "title": "Difficulty"
          },
          "scheduled_for": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled For"
          }
        },
        "type": "object",
        "required": [
          "senior_member_id",
          "voice_member_id",
          "scenario",
          "difficulty"
        ],
        "title": "PracticeCallIn"
      },
      "PracticeCallOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "circle_id": {
            "type": "string",
            "title": "Circle Id"
          },
          "senior_member_id": {
            "type": "string",
            "title": "Senior Member Id"
          },
          "voice_member_id": {
            "type": "string",
            "title": "Voice Member Id"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "feedback": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/TranscriptLine"
            },
            "type": "array",
            "title": "Transcript"
          },
          "turns": {
            "type": "integer",
            "title": "Turns"
          },
          "safety_stop": {
            "type": "boolean",
            "title": "Safety Stop"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "scheduled_for": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled For"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "circle_id",
          "senior_member_id",
          "voice_member_id",
          "scenario",
          "difficulty",
          "status",
          "outcome",
          "score",
          "feedback",
          "transcript",
          "turns",
          "safety_stop",
          "duration_seconds",
          "scheduled_for",
          "started_at",
          "ended_at",
          "created_at"
        ],
        "title": "PracticeCallOut"
      },
      "ScenarioOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Difficulties"
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "description",
          "difficulties"
        ],
        "title": "ScenarioOut"
      },
...
      "TranscriptLine": {
        "properties": {
          "speaker": {
            "type": "string",
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "t_ms": {
            "type": "integer",
            "title": "T Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "speaker",
          "text"
        ],
        "title": "TranscriptLine"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}